refactor: reader, writer -> adaptor로 통합
