--- backend/src/main/java/kernel360/techpick/feature/infrastructure/tag/writer/TagWriter.java
@@ -1,10 +1,0 @@
-package kernel360.techpick.feature.infrastructure.tag.writer;
-
-import kernel360.techpick.core.model.tag.Tag;
-
-public interface TagWriter {
-
-	Tag writeTag(Tag tag);
-
-	void removeTag(Long tagId);
-}
